+++ CHANGELOG.md
@@ -98,6 +98,7 @@ All notable changes to this project will be documented in this file.
 - Hex View - incorrect table height when switching tags
 - [#2525] AMF editor (like in Sol cookie editor) - saving empty standard array `[]`
 - AS3 - names with custom namespaces not being deobfuscated
+- AS Debugger - order of trace commands
 
  ### Changed
 - Icon of "Deobfuscation options" menu from pile of pills to medkit